Refactor - move useWindowSize as standalone - move useMatchMedia as standalone - move fluidSizing as standalone - update main artifak pkg exports - renamed createBaseComponents to createComponents - add generation of base component to result of generated comps for createComponents
