+++ packages/fluidsizing/src/index.ts
@@ -1,0 +1,1 @@
+export { fluidSizing } from './fluidSizing';
