+++ packages/media/src/index.ts
@@ -6,8 +6,6 @@ export { MediaInputQueries } from './mediaInput';
 export { MediaTypes } from './mediaTypes';
 export { MediaAccessibilityQueries } from './mediaAccessibilty';
 export { MediaDisplayQueries } from './mediaDisplay';
-export { useMatchMedia } from './useMatchMedia';
-export { useWindowSize } from './useWindowSize';
 
 export {
   all,
