+++ packages/typography/__tests__/typography.test.ts
@@ -1,11 +1,5 @@
 import { position } from 'styled-system';
-
-import {
-  createTypographyComponents,
-  getStyleProps,
-  fluidSizing,
-  fontWeight
-} from '../src';
+import { createTypographyComponents, getStyleProps, fontWeight } from '../src';
 
 describe('@artifak/typography', () => {
   describe('createTypographyComponents - generates typography React components based on styles object argument provided', () => {
@@ -67,25 +61,6 @@ describe('@artifak/typography', () => {
         element: 'div',
         styleProps: [position]
       });
-    });
-  });
-
-  describe('fluidSizing - writes sizing CSS rule that dynamically adapts to screen width based on provided min/max constraits', () => {
-    it('should return FALSE by default', () => {
-      const expected = '';
-      expect(
-        fluidSizing(void 0 as any, void 0 as any, void 0 as any, void 0 as any)
-      ).toEqual(expected);
-    });
-
-    it('should return FALSE by when provided falsy arguments', () => {
-      const expected = '';
-      expect(fluidSizing(void 0 as any, 96, 300, 1200)).toEqual(expected);
-    });
-
-    it('should return a CSS string rule', () => {
-      const expected = `calc(48px + (96 - 48) * ((100vw - 300px) / (1200 - 300)))`;
-      expect(fluidSizing(48, 96, 300, 1200)).toEqual(expected);
     });
   });
