+++ packages/typography/src/index.tsx
@@ -4,5 +4,4 @@ export {
 } from './createTypographyComponents';
 export { TypographyBase } from './TypographyBase';
 export { fontWeight } from './fontWeight';
-export { fluidSizing } from './fluidSizing';
 export { TypographyBaseProps } from './typings';
