+++ packages/usematchmedia/__mocks__/useMatchMedia.mock.ts
@@ -1,0 +1,18 @@
+import mediaQuery from 'css-mediaquery';
+
+Object.defineProperty(window, 'matchMedia', {
+  writable: true,
+  value: jest.fn().mockImplementation(query => ({
+    matches: mediaQuery.match(query, {
+      width: window.innerWidth,
+      height: window.innerHeight
+    }),
+    media: query,
+    onchange: null,
+    addListener: jest.fn(), // Deprecated
+    removeListener: jest.fn(), // Deprecated
+    addEventListener: jest.fn(),
+    removeEventListener: jest.fn(),
+    dispatchEvent: jest.fn()
+  }))
+});
