+++ packages/usematchmedia/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "@artifak/usematchmedia",
+  "version": "1.0.0",
+  "description": "A React hook for matching media queries.",
+  "keywords": [],
+  "author": "Julian Low",
+  "license": "ISC",
+  "sideEffects": false,
+  "main": "dist/usematchmedia.cjs.js",
+  "module": "dist/usematchmedia.esm.js",
+  "src": "src/index.ts",
+  "directories": {
+    "lib": "src",
+    "test": "__tests__"
+  },
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/heyjul3s/artifak.git"
+  },
+  "scripts": {
+    "build": "rimraf dist && builder",
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "bugs": {
+    "url": "https://github.com/heyjul3s/artifak/issues"
+  },
+  "homepage": "https://github.com/heyjul3s/artifak#readme",
+  "devDependencies": {
+    "@artifak/bundler": "^1.1.3",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1"
+  },
+  "peerDependencies": {
+    "react": ">=17.0.1",
+    "react-dom": ">=17.0.1"
+  }
+}
