+++ packages/usematchmedia/src/index.ts
@@ -1,0 +1,1 @@
+export { useMatchMedia } from './useMatchMedia';
