+++ packages/usewindowsize/src/index.ts
@@ -1,0 +1,2 @@
+export { useWindowSize } from './useWindowSize';
+export { WindowSize } from './typings';
