+++ packages/usewindowsize/src/typings.ts
@@ -1,0 +1,6 @@
+export type WindowSize = {
+  innerHeight: number;
+  innerWidth: number;
+  outerHeight: number;
+  outerWidth: number;
+};
